move wrappers to index
