+++ shared/profile/wot-author/index.tsx
@@ -8,6 +8,7 @@ import * as RouteTreeGen from '../../actions/route-tree-gen'
 import * as Tracker2Types from '../../constants/types/tracker2'
 import {SiteIcon} from '../../profile/generic/shared'
 import * as ProfileGen from '../../actions/profile-gen'
+import * as Tracker2Constants from '../../constants/tracker2'
 
 // PICNIC-1059 Keep in sync with server limit (yet to be implemented)
 const statementLimit = 700
@@ -56,6 +57,91 @@ export type Proof = {
 type Checkboxed = {
   checked: boolean
   onCheck: (_: boolean) => void
+}
+
+export const Question1Wrapper = (props: Container.RouteProps<{username: string}>) => {
+  const voucheeUsername = Container.getRouteProps(props, 'username', '')
+  const nav = Container.useSafeNavigation()
+  const dispatch = Container.useDispatch()
+  let error = Container.useSelector(state => state.profile.wotAuthorError)
+  if (!error && !voucheeUsername) {
+    error = 'Routing missing username.'
+  }
+  const {username: trackerUsername, assertions} = Container.useSelector(state =>
+    Tracker2Constants.getDetails(state, voucheeUsername)
+  )
+  let proofs: Proof[] = []
+  if (trackerUsername === voucheeUsername) {
+    if (assertions) {
+      proofs = Array.from(assertions, ([_, assertion]) => assertion).filter(x => x.type !== 'stellar')
+    }
+  } else {
+    error = `Proofs not loaded: ${trackerUsername} != ${voucheeUsername}`
+  }
+  const onSubmit = (answer: Question1Answer) => {
+    dispatch(
+      nav.safeNavigateAppendPayload({
+        path: [{props: {question1Answer: answer, username: voucheeUsername}, selected: 'profileWotAuthorQ2'}],
+      })
+    )
+  }
+  return (
+    <Question1
+      error={error}
+      initialVerificationType={'in_person'}
+      onSubmit={onSubmit}
+      proofs={proofs}
+      voucheeUsername={voucheeUsername}
+    />
+  )
+}
+
+export const Question2Wrapper = (
+  props: Container.RouteProps<{
+    username: string
+    question1Answer: Question1Answer
+  }>
+) => {
+  const voucheeUsername = Container.getRouteProps(props, 'username', '')
+  const question1Answer = Container.getRoutePropsOr(props, 'question1Answer', 'error')
+  const nav = Container.useSafeNavigation()
+  const dispatch = Container.useDispatch()
+  let error = Container.useSelector(state => state.profile.wotAuthorError)
+  if (!error && !voucheeUsername) {
+    error = 'Routing missing username.'
+  }
+  if (!error && question1Answer === 'error') {
+    error = 'Routing missing q1 answer.'
+  }
+  const waiting = Container.useAnyWaiting(Constants.wotAuthorWaitingKey)
+  const onSubmit = ({statement}: {statement: string}) => {
+    if (question1Answer === 'error') {
+      return
+    }
+    const {otherText, verificationType} = question1Answer
+    dispatch(
+      ProfileGen.createWotVouch({
+        otherText,
+        proofs: [], // PICNIC-1087 TODO
+        statement,
+        username: voucheeUsername,
+        verificationType,
+      })
+    )
+  }
+  const onBack = () => {
+    dispatch(ProfileGen.createWotVouchSetError({error: ''}))
+    dispatch(nav.safeNavigateUpPayload())
+  }
+  return (
+    <Question2
+      error={error}
+      onBack={onBack}
+      onSubmit={onSubmit}
+      voucheeUsername={voucheeUsername}
+      waiting={waiting}
+    />
+  )
 }
 
 export const Question1 = (props: Question1Props) => {
